% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empCop.R
\docType{class}
\name{cbCopula-class}
\alias{cbCopula-class}
\alias{.cbCopula}
\title{Checkerboard copula class}
\value{
A cbCopula object, which can be use to do... nothing yet !
}
\description{
Checkerboard copula class
}
\section{Slots}{

\describe{
\item{\code{m}}{numeric : checkerboard parameter defining the "grid" dividing the space.}
}}

