% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbCopula.R
\name{cbCopula}
\alias{cbCopula}
\title{cbCopula contructor}
\usage{
cbCopula(x, m = nrow(x), pseudo = FALSE)
}
\arguments{
\item{x}{the data to be used}

\item{m}{checkerboard parameter}

\item{pseudo}{Boolean, defaults to `FALSE`. Set to `TRUE` if you are already
providing pseudo datas into the `x` argument.}
}
\value{
a cbCopula object
}
\description{
cbCopula contructor
}
