% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empCop.R
\docType{class}
\name{cbkmCopula-class}
\alias{cbkmCopula-class}
\alias{.cbkmCopula}
\title{Checkerboard with known margins copula class}
\value{
A cbCopula object, which can be use to do... nothing yet !
}
\description{
Checkerboard with known margins copula class
}
\section{Slots}{

\describe{
\item{\code{m}}{numeric : checkerboard parameter defining the "grid" dividing the space.}

\item{\code{margins}}{numeric integers refering to the margins you want to assign the known_cop to}

\item{\code{known_cop}}{Copula a copula object representing the known copula for the selected margins.}
}}

