% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empCop.R
\name{cbkmCopula}
\alias{cbkmCopula}
\title{cbkmCopula contructor}
\usage{
cbkmCopula(x, m = nrow(x), pseudo = FALSE, margins_numbers = NULL,
  known_cop = NULL)
}
\arguments{
\item{x}{the data to be used}

\item{m}{checkerboard parameter}

\item{pseudo}{Boolean, defaults to `FALSE`. Set to `TRUE` if you are already providing pseudo datas into the `x` argument.}

\item{margins_numbers}{numeric integers refering to the margins you want to assign the known_cop to}

\item{known_cop}{Copula a copula object representing the known copula for the selected margins.}
}
\value{
a cbCopula object
}
\description{
cbkmCopula contructor
}
