% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvexCombCopula.R
\docType{class}
\name{ConvexCombCopula-class}
\alias{ConvexCombCopula-class}
\alias{.ConvexCombCopula}
\title{ConvexCombCopula class}
\value{
a ConvexCombCopula object.
}
\description{
ConvexCombCopula class
}
\section{Slots}{

\describe{
\item{\code{copulas}}{list of copulas of same dimension}

\item{\code{alpha}}{numeric. A vector of (positive) weights.}
}}

