% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empCop.R
\name{ConvexCombCopula}
\alias{ConvexCombCopula}
\title{ConvexCombCopula}
\usage{
ConvexCombCopula(copulas, alpha = rep(1, length(copulas)))
}
\arguments{
\item{copulas}{a list of copulas of same dimention}

\item{alpha}{a vector of (positive) weights

The convexcombcopula class is used to build convex combinations of copulas,
with given positives weights. The rCopula and pCopula functions works for
thoose copulas, assuming they work for the given copulas that we combined
in a convex way.}
}
\value{
a ConvexCombCopula object
}
\description{
ConvexCombCopula
}
