% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalCopula.R
\docType{class}
\name{empiricalCopula-class}
\alias{empiricalCopula-class}
\title{Empirical Copula class (virtual mother class)}
\description{
Empirical Copula class (virtual mother class)
}
\section{Slots}{

\describe{
\item{\code{pseudo_data}}{matrix : pseudo_data that the empirical copula is based on.}
}}

