% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empCop.R
\name{vCopula}
\alias{vCopula}
\title{Copula volume on hyper-boxes}
\usage{
vCopula(u, v, copula)
}
\arguments{
\item{u}{numeric minimum point of the hyper-rectangle}

\item{v}{numeric maximum point of the hyper-rectangle}

\item{copula}{the copula to calcule it's measure on [u,v]

u must be piecewise smaller than v, otherwise the function will return an error.

This function calculates the measure of the copula according to the algorythme proposed by :
Umberto Cherubini & Silvia Romagnoli (2009) Computing the
Volume of n-Dimensional Copulas, Applied Mathematical Finance, 16:4, 307-314, DOI:
  10.1080/13504860802597311 link : http://dx.doi.org/10.1080/13504860802597311}
}
\value{
the measure of the copula
}
\description{
Copula volume on hyper-boxes
}
\examples{
cop = copula::archmCopula("Clayton",0.7,3)
vCopula(rep(0,3),rep(1,3),cop)

}
