% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbkmCopula.R
\name{cbkmCopula}
\alias{cbkmCopula}
\title{cbkmCopula contructor}
\usage{
cbkmCopula(x, m = nrow(x), pseudo = FALSE, margins_numbers = NULL,
  known_cop = NULL, quiet = FALSE, exception_type = "checkerboard")
}
\arguments{
\item{x}{the data to be used}

\item{m}{checkerboard parameter}

\item{pseudo}{Boolean, defaults to `FALSE`. Set to `TRUE` if you are already providing pseudo datas into the `x` argument.}

\item{margins_numbers}{numeric integers refering to the margins you want to assign the known_cop to}

\item{known_cop}{Copula a copula object representing the known copula for the selected margins.}

\item{quiet}{Should the constructor be verbose or not ?}

\item{exception_type}{can be one of "checkerboard" or "indep". Condition the constrution for checkerboard margins when not points are avaliable.}
}
\value{
a cbCopula object
}
\description{
Given some empirical data, and given some knwon copula estimation on a sub-vector of this data,
the checkerboar with known margins construction consist in
a conditional pattern where the checkerboar part is conditional on the known part of the copula
This allows for high-dimensional-aware contructions.
}
\examples{
true_copula <- onacopulaL(family = 'Clayton',
nacList = list(iTau(getAcop('Clayton'), 0.6), 1:4))

dataset <- rCopula(50,true_copula)

known_margins <- c(2,3)
known_clayton <- onacopulaL(
  family = 'Clayton',
  nacList = list(iTau(getAcop('Clayton'), 0.6), 1:2)
)

cop <- cbkmCopula(x = dataset,
                  m = 5,
                  pseudo = TRUE,
                  margins_numbers = known_margins,
                  known_cop = known_clayton)


u=rbind(rep(0,4),matrix(rep(0.7,12),nrow=3),rep(1,4))

pCopula(u,cop)
}
