% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvexCombCopula.R
\name{ConvexCombCopula}
\alias{ConvexCombCopula}
\title{ConvexCombCopula class}
\usage{
ConvexCombCopula(copulas, alpha = rep(1, length(copulas)))
}
\arguments{
\item{copulas}{a list of copulas of same dimention}

\item{alpha}{a vector of (positive) weights}
}
\value{
a ConvexCombCopula object
}
\description{
The ConvexCombcopula class is used to build convex combinations of copulas,
with given positives weights. The rCopula and pCopula functions works for
thoose copulas, assuming they work for the given copulas that we combined
in a convex way.
}
\examples{
copulas <- list(
  copula::archmCopula('gumbel',3,dim=2),
  copula::archmCopula('clayton',-1,dim=2)
)
alpha <- c(1,4)

(cop <- ConvexCombCopula(copulas,alpha))

plot(rCopula(100,cop))
pCopula(c(0.5,0.7),cop)
}
